move subscription list component, fix clear area selection
